Make BackgroundQueue a top-level class
